Split security config
