+++ management-app-main/src/main/java/com/teamcore/manageapp/main/SiteApplication.java
@@ -1,9 +1,12 @@
 package com.teamcore.manageapp.main;
 
+import com.teamcore.manageapp.main.config.AppConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackageClasses = {AppConfig.class})
 public class SiteApplication {
 
 	public static void main(String[] args) {
