+++ management-app-service/src/main/java/com/teamcore/manageapp/service/config/ServiceConfigsMarker.java
@@ -1,0 +1,3 @@
+package com.teamcore.manageapp.service.config;
+
+public interface ServiceConfigsMarker { }
