+++ management-app-service/src/main/java/com/teamcore/manageapp/service/config/ServiceSecurityConfig.java
@@ -3,22 +3,14 @@ package com.teamcore.manageapp.service.config;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
-import java.io.IOException;
 
 @Configuration
 @EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+public class ServiceSecurityConfig extends WebSecurityConfigurerAdapter {
 
     private DataSource dataSource;
 
@@ -40,19 +32,4 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .authoritiesByUsernameQuery(authoritiesByUsernameQuery);
     }
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.authorizeRequests().antMatchers("/").authenticated().anyRequest().permitAll()
-        .and()
-        .formLogin()
-        .loginProcessingUrl("/login").loginPage("/login")
-        .failureHandler((httpServletRequest, httpServletResponse, e) -> {
-            HttpSession session = httpServletRequest.getSession();
-            session.setAttribute("authFail", "true");
-            httpServletResponse.sendRedirect("/login");
-//            httpServletResponse.addHeader("authFail", "true");
-//            httpServletRequest.getRequestDispatcher("/login").forward(httpServletRequest, httpServletResponse);
-        })
-        .permitAll();
-    }
 }
