+++ management-app-service/src/main/java/com/teamcore/manageapp/service/dao/DaoMarker.java
@@ -1,0 +1,3 @@
+package com.teamcore.manageapp.service.dao;
+
+public interface DaoMarker { }
