+++ management-app-service/src/test/java/com/teamcore/manageapp/service/config/TestConfig.java
@@ -1,7 +1,7 @@
 package com.teamcore.manageapp.service.config;
 
-import com.teamcore.manageapp.service.TestFactory;
-import com.teamcore.manageapp.service.dao.user.UserDAOImpl;
+import com.teamcore.manageapp.service.dao.DaoMarker;
+import com.teamcore.manageapp.service.utils.TestUtilsMarker;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -10,8 +10,7 @@ import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import javax.sql.DataSource;
 
 @Configuration
-//TODO create marker interfaces!!!!
-@ComponentScan(basePackageClasses = {UserDAOImpl.class, TestFactory.class, DataBaseConfig.class })
+@ComponentScan(basePackageClasses = {DaoMarker.class, TestUtilsMarker.class, ServiceConfigsMarker.class })
 public class TestConfig {
 
    @Bean
