+++ management-app-service/src/test/java/com/teamcore/manageapp/service/utils/TestUtilsMarker.java
@@ -1,0 +1,3 @@
+package com.teamcore.manageapp.service.utils;
+
+public interface TestUtilsMarker { }
