+++ management-app-web/pom.xml
@@ -13,6 +13,19 @@
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
 
+	<dependencies>
+		<dependency>
+			<groupId>com.teamcore</groupId>
+			<artifactId>management-app-service</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-thymeleaf</artifactId>
+        </dependency>
+
+    </dependencies>
 
 	<build>
 		<finalName>${project.artifactId}-${project.version}</finalName>
