+++ management-app-web/src/main/java/com/teamcore/manageapp/web/config/WebConfigsMarker.java
@@ -1,0 +1,3 @@
+package com.teamcore.manageapp.web.config;
+
+public interface WebConfigsMarker { }
