+++ management-app-web/src/main/java/com/teamcore/manageapp/web/controllers/ControllersMarker.java
@@ -1,0 +1,3 @@
+package com.teamcore.manageapp.web.controllers;
+
+public interface ControllersMarker {}
