+++ management-app-web/src/main/java/com/teamcore/manageapp/web/controllers/IndexController.java
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class IndexController {
     @RequestMapping("/")
     public String index() {
-        return "index";
+        return "login";
     }
 }
